+++ src/test/scala/satify/problems/NQueensTest.scala
@@ -34,3 +34,12 @@ class NQueensTest extends AnyFlatSpec with Matchers:
     println("NQueens 10x10")
     printNQueensFromDimacs(10, sol.assignment.head.parModel)
   }
+
+  /*"NQueens 15x15" should "be SAT" in {
+    val oCnf = DimacsCNF.read("src/main/resources/cnf/nqueens20.txt")
+    oCnf should matchPattern { case Some(_) => }
+    val sol = Solver(DPLL).solve(oCnf.get)
+    sol should matchPattern { case Solution(SAT, _) => }
+    println("NQueens 15x15")
+    printNQueensFromDimacs(15, sol.assignment.head.parModel)
+  }*/
