+++ src/test/scala/satify/update/solver/dpll/DpllTest.scala
@@ -10,7 +10,7 @@ import satify.model.dpll.OrderedSeq.*
 import satify.model.dpll.{Decision, DecisionTree}
 import satify.model.{CNF, Variable}
 import satify.update.solver.dpll.Dpll.*
-import satify.update.solver.dpll.Dpll.extractSolutions
+import satify.update.solver.dpll.DpllUtils.extractSolutions
 import satify.update.solver.dpll.PartialModelUtils.*
 
 class DpllTest extends AnyFlatSpec with Matchers:
