+++ src/test/scala/satify/update/solver/dpll/PartialModelUtilsTest.scala
@@ -6,7 +6,7 @@ import satify.model.CNF.*
 import satify.model.dpll.OrderedSeq.{given_Ordering_Variable, seq}
 import satify.model.dpll.{Constraint, Decision, PartialModel}
 import satify.model.{CNF, Variable}
-import satify.update.solver.dpll.DPLL.dpll
+import satify.update.solver.dpll.Dpll.dpll
 import satify.update.solver.dpll.PartialModelUtils.*
 
 class PartialModelUtilsTest extends AnyFlatSpec with Matchers:
